Updated to send mail and float number

+++ app/purchase_tracker/models.py
@@ -9,7 +9,7 @@ class PurchaseTrackerAccount(db.Model):
     subject = db.Column(db.String(255), nullable=False, info={'label': u"ชื่อเรื่อง"})
     number = db.Column(db.String(255), nullable=False, info={'label': u"เลขที่หนังสือ"})
     booking_date = db.Column('booking_date', db.Date(), nullable=False, info={'label': u'วันที่หนังสือ'})
-    amount = db.Column('amount', db.Integer(), info={'label': u'วงเงินหลักการ'})
+    amount = db.Column('amount', db.Float(), info={'label': u'วงเงินหลักการ'})
     formats = db.Column('formats', db.String(255), info={'label': u'รูปแบบหลักการ'})
     creation_date = db.Column('creation_datetime', db.DateTime(timezone=True), nullable=False)
     staff_id = db.Column('staff_id', db.ForeignKey('staff_account.id'), nullable=False)
